+++ src/task/set_network_threshold.go
@@ -3,7 +3,7 @@ package task
 import (
 	"log"
 	"github.com/project-nano/framework"
-	"service"
+	"github.com/project-nano/cell/service"
 	"fmt"
 )